[OPIK-1415] global user state
