+++ apps/opik-frontend/src/components/pages-shared/onboarding/ApiKeyCard/ApiKeyCard.tsx
@@ -1,14 +1,19 @@
-import React from "react";
-import usePluginsStore from "@/store/PluginsStore";
+import ApiKeyInput from "@/components/shared/ApiKeyInput/ApiKeyInput";
+import { useActiveUserApiKey } from "@/store/AppStore";
 
 const ApiKeyCard = () => {
-  const ApiKeyCard = usePluginsStore((state) => state.ApiKeyCard);
+  const apiKey = useActiveUserApiKey();
 
-  if (ApiKeyCard) {
-    return <ApiKeyCard />;
+  if (!apiKey) {
+    return null;
   }
 
-  return null;
+  return (
+    <div className="flex flex-1 flex-col justify-between gap-4 rounded-md border bg-white p-6">
+      <div className="comet-title-xs text-foreground-secondary">API key</div>
+      <ApiKeyInput apiKey={apiKey} />
+    </div>
+  );
 };
 
 export default ApiKeyCard;
